+++ src/updater/check_update.h
@@ -46,14 +46,14 @@ namespace updater {
 
     // Returns the number of days that this client should wait for the
     // next "check for updates".
-    int getWaitDays() const { return m_waitDays; }
+    double getWaitDays() const { return m_waitDays; }
 
   private:
     Type m_type;
     base::Version m_version;
     std::string m_url;
     Uuid m_uuid;
-    int m_waitDays;
+    double m_waitDays;
   };
 
   // Delegate called by CheckUpdate when the request to the server is
